stringifiable base class
