+++ Caching.Memory/Library/Generic/IGenericMemoryCache.cs
@@ -3,11 +3,9 @@
 namespace Woody230.Caching.Memory;
 
 /// <summary>
-/// Represents a memory cache that is typed.
+/// Represents a memory cache that is generically typed.
 /// </summary>
-/// <typeparam name="TKey">The type of key.</typeparam>
-/// <typeparam name="TValue">The type of value.</typeparam>
-public interface IGenericMemoryCache<TKey, TValue>
+public interface IGenericMemoryCache<TKey, TValue> where TKey: notnull
 {
     /// <summary>
     /// The keys in the cache.
